requirements.txt
# Core
numpy>=1.23.0
torch>=2.1.0
torch-geometric>=2.4.0
transformers>=4.40.0
tqdm>=4.66.0

# Data / config
pyyaml>=6.0
pandas>=1.5.0

# Training / clustering
scikit-learn>=1.2.0
scipy>=1.10.0

# Optional text generation via API
openai>=1.0.0
